requirements.txt
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
filelock @ file:///croot/filelock_1700591183607/work
fonttools @ file:///croot/fonttools_1713551344105/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp310-cp310-linux_x86_64.whl#sha256=a2434a313a5eb5b4fdbec6c770abfde859f643eff7b3a703791bf8b73e54c394
pillow @ file:///croot/pillow_1721059439630/work
ply==3.11
pyparsing @ file:///croot/pyparsing_1725041624202/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1721410876868/work
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///croot/sympy_1724938189289/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///croot/tqdm_1724853939799/work
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1718912636303/work
